First phase of spliting core.py
Here is what we have

robot@evb[ev3dev-lang-python]# tree ev3dev/
ev3dev/
├── button.py
├── control
│   ├── GyroBalancer.py
│   ├── __init__.py
│   └── rc_tank.py
├── display.py
├── fonts
//
// chopped most of fonts to limit the output here
//
│   ├── timR18.pbm
│   ├── timR18.pil
│   ├── timR24.pbm
│   └── timR24.pil
├── helper.py
├── __init__.py
├── led.py
├── motor.py
├── _platform
│   ├── brickpi.py
│   ├── ev3.py
│   ├── evb.py
│   ├── __init__.py
│   └── pistorms.py
├── port.py
├── sensor
│   ├── __init__.py
│   └── lego.py
├── sound.py
├── version.py
└── webserver.py

4 directories, 421 files
robot@evb[ev3dev-lang-python]#
